fix: removed warnings in IntegrationTests

+++ test/Chirp.Web.Tests/IntegrationTests.cs
@@ -1,14 +1,7 @@
-﻿using System.ComponentModel;
-using System.Net;
-using Chirp.Core;
-using Chirp.Web;
+﻿using Chirp.Core;
 using Chirp.Infrastructure.Chirp.Repositories;
 using Chirp.Infrastructure.Chirp.Services;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.VisualStudio.TestPlatform.TestHost;
-using NUnit.Framework;
 using Xunit;
-using Xunit.Abstractions;
 using Assert = Xunit.Assert;
 
 namespace Chirp.Web.Tests;
@@ -17,14 +10,14 @@ namespace Chirp.Web.Tests;
 public class IntegrationTests : IAsyncLifetime
 {
 
-    private IAuthorRepository _authorrepo;
-    private ICheepRepository _cheeprepo;
-    private IFollowRepository _followrepo;
+    private IAuthorRepository? _authorrepo;
+    private ICheepRepository? _cheeprepo;
+    private IFollowRepository? _followrepo;
 
-    private TestUtilities _utils;
-    private CheepDbContext _context;
+    private TestUtilities? _utils;
+    private CheepDbContext? _context;
     
-    private IChirpService _service;
+    private IChirpService? _service;
     
     public async Task InitializeAsync()
     {
@@ -47,6 +40,10 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestAddCheep()
     {
+        if (_utils == null || _cheeprepo == null || _authorrepo == null)
+        {
+            return;
+        }
         var cheepsBefore = await _cheeprepo.GetCheepsFromAuthor(0, "Mellie Yost");
         
         var author = await _authorrepo.GetAuthorByName("Mellie Yost");
@@ -69,6 +66,10 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task AddCheepchirpServiceNonExistingAuthor()
     {
+        if (_service == null || _utils == null || _authorrepo == null)
+        {
+            return;
+        }
         await _service.AddCheep("testest", "NewAuthor", "@newauthor.com");
     
         var author = await _authorrepo.GetAuthorByName("NewAuthor");
@@ -84,7 +85,11 @@ public class IntegrationTests : IAsyncLifetime
     {
         //We first make a user follow another user. Show that the cheeps should now be the sum of their cheeps in their private timeline
         //Octavio Wagganer(15) follow Mellie Yost(7)
-
+        if (_service == null)
+        {
+            return;
+        }
+        
         string authorname1 = "Octavio Wagganer";
         string authorname2 = "Mellie Yost";
 
@@ -99,6 +104,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestCheepHasFollowIfAuthorFollows()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         string authorname1 = "Octavio Wagganer";
         string authorname2 = "Mellie Yost";
 
@@ -114,6 +124,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestDeleteFollows()
     {
+        if (_service == null || _authorrepo == null || _followrepo == null)
+        {
+            return;
+        }
+        
         await _authorrepo.CreateAuthor("erik", "hahaemail@gmail.com");
         await _authorrepo.CreateAuthor("lars", "bahaemail@gmail.com");
         await _followrepo.AddFollowing("erik", "lars");
@@ -131,6 +146,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestCountingOfLikesOnCheep()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         await _service.AddLike("Octavio Wagganer", 1);
         await _service.AddLike("Mellie Yost", 1);
 
@@ -143,6 +163,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task CanRemoveLike()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         await _service.AddLike("Octavio Wagganer", 1);
         var likes1Amount = await _service.CountLikes(1);
         await _service.RemoveLike("Octavio Wagganer", 1);
@@ -155,6 +180,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task CanRemoveLikeData()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         string author = "Octavio Wagganer";
         await _service.AddLike(author, 1);
         var likes1Amount = await _service.CountLikes(1);
